Remove all circular imports identified by Rollup
