+++ src/lib/containers/access_requirement_list/AcceptedRequirements.tsx
@@ -15,7 +15,7 @@ import {
 } from '../../utils/synapseTypes/'
 import { SynapseClient } from '../../utils'
 import AccessApprovalCheckMark from './AccessApprovalCheckMark'
-import { SUPPORTED_ACCESS_REQUIREMENTS } from './AccessRequirementList'
+import { SUPPORTED_ACCESS_REQUIREMENTS } from './SupportedAccessRequirements'
 
 export type AcceptedRequirementsProps = {
   user: UserProfile | undefined
