+++ src/lib/containers/access_requirement_list/SupportedAccessRequirements.ts
@@ -1,0 +1,6 @@
+export enum SUPPORTED_ACCESS_REQUIREMENTS {
+  SelfSignAccessRequirement = 'org.sagebionetworks.repo.model.SelfSignAccessRequirement',
+  TermsOfUseAccessRequirement = 'org.sagebionetworks.repo.model.TermsOfUseAccessRequirement',
+  ManagedACTAccessRequirement = 'org.sagebionetworks.repo.model.ManagedACTAccessRequirement',
+  ACTAccessRequirement = 'org.sagebionetworks.repo.model.ACTAccessRequirement',
+}
